Solves address parsing error by avoiding BigNumber
When 'ownerStorage' is an address starting with 0x00 the BigNumber library rounds down the address to 38 characters, thus making it an invalid address and causing getAddress to throw a fatal error.   This commit avoids the use of BigNumber and just uses the last 20 bytes to solve this rare, edge case problem.